reafactor useIsMobile to useDeviceBreakpoints, include isTablet
