--- packages/react/src/hooks/useIsMobile.ts
@@ -1,28 +1,0 @@
-import { useState, useEffect } from 'react';
-
-const mobileBreakpoint = 768;
-const useIsMobile = () => {
-  const [isMobile, setIsMobile] = useState(false);
-
-  const handleResize = () => {
-    if (window.innerWidth < mobileBreakpoint) {
-      setIsMobile(true);
-    } else {
-      setIsMobile(false);
-    }
-  };
-
-  useEffect(() => {
-    handleResize();
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
-
-  return isMobile;
-};
-
-export default useIsMobile;
